Add files via upload
CNN Classification- Lego Bricks

The data set contains rendered images of 16 different types of Lego bricks. This is an image classification task: build a model that can correctly identify lego bricks. Used the train folder and build a model to predict the category of each image. Then, validated the model on the images in the valid folder.